requirements.txt
fastapi
mangum
azure-storage-blob